// src/types/completeSetupMfaCommandMethod.ts
/**
 * Generated by orval v8.5.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type CompleteSetupMfaCommandMethod = typeof CompleteSetupMfaCommandMethod[keyof typeof CompleteSetupMfaCommandMethod];


export const CompleteSetupMfaCommandMethod = {
  TOTP: 'TOTP',
  EMAIL: 'EMAIL',
  WEBAUTHN: 'WEBAUTHN',
} as const;